+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),  
 and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.10.3.1] - 2025-01-07
-### Bugfix simbriefID
+### Bugfix
+- simbriefID null value
 
 ## [0.10.3] - 2025-01-07
 ### Fixed